Switch to mingw-w64
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1883 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/c3270/c3270.c
@@ -1621,8 +1621,8 @@ start_auto_shortcut(void)
     sprintf(delenv, "%s=%s", DELENV, linkpath);
     putenv(delenv);
     h = ShellExecute(NULL, "open", linkpath, "", tempdir, SW_SHOW);
-    if ((int)h <= 32) {
-	fprintf(stderr, "ShellExecute failed, error %d\n", (int)h);
+    if ((uintptr_t)h <= 32) {
+	fprintf(stderr, "ShellExecute failed, error %d\n", (int)(uintptr_t)h);
 	x3270_exit(1);
     }
